Code cleaning / minor style edits
