+++ atomai/losses_metrics/__init__.py
@@ -1,2 +1,4 @@
-from .losses import *
-from .metrics import *
+from .losses import focal_loss, dice_loss
+from .metrics import IoU
+
+__all__ = ['focal_loss', 'dice_loss', 'IoU']
